+++ luscii_patient_actions_sdk_platform_interface/lib/model/luscii_environment.dart
@@ -1,0 +1,11 @@
+/// The environment for the SDK.
+enum LusciiEnvironment {
+  /// The production environment.
+  production,
+
+  /// The acceptance environment.
+  acceptance,
+
+  /// The test environment.
+  test,
+}